fix: change expires_in type from int to float in ResponceToken class

+++ backend/app/schemas/telegram_auth.py
@@ -36,5 +36,5 @@ class RequestInitData(BaseModel):
 
 class ResponceToken(BaseModel):
     access_token: str
-    expires_in: int
+    expires_in: float
     user_id: int
